+++ mne/inverse_sparse/tests/test_mxne_optim.py
@@ -111,9 +111,10 @@ def test_tf_mxne():
 
     M, G, active_set = _generate_tf_data()
 
-    X_hat_tf, active_set_hat_tf, E = tf_mixed_norm_solver(
+    X_hat_tf, active_set_hat_tf, E, gap_tfmxne = tf_mixed_norm_solver(
         M, G, alpha_space, alpha_time, maxit=200, tol=1e-8, verbose=True,
-        n_orient=1, tstep=4, wsize=32)
+        n_orient=1, tstep=4, wsize=32, return_gap=True)
+    assert_array_less(gap_tfmxne, 1e-8)
     assert_array_equal(np.where(active_set_hat_tf)[0], active_set)